requirements.txt
# Core web app
streamlit
python-dotenv

# Data & computation
pandas
numpy
scikit-learn
scipy

# Visualization
plotly

# PDF/report generation
reportlab

# LLM integration
google-generativeai

# Utilities
pydantic
requests

# Optional: caching & performance
cachetools

# Dev tools (optional)
black
isort
flake8